LPD-61412 Move useBatchEngineExportTask to a hook
